fix: Use `carousel_slider_plus` instead of `carousel_slider` package
Temporary fix for the issue with carousel_slider package. The package is
not working with the latest flutter version (due to name conflict with
CarouselController in material.dart). The carousel_slider_plus package is
a fork of the carousel_slider package with the necessary changes to
resolve the issue.

Future plan is to implement carousel without using any package.